gets

+++ kernel/include/lib/bareio.h
@@ -16,6 +16,6 @@ void ksprintf(byte*, const char*, ...);
 void krprintf(const char*, ...);
 void vkrprintf(const char*, va_list);
 void kuprintf(const char*, uint32_t);
-uint16_t get_line(char*, uint16_t);
+uint32_t get_line(char*, uint32_t);
 
 #endif
